requirements.txt
py-cord==2.3.2
requests~=2.28.1
asgiref==3.6.0
nest-asyncio==1.5.6
pillow~=9.3.0
numpy==1.21.0
pandas==1.3.3
python-dotenv==0.21.1
# yfinance
# google-search-results
# stability-sdk
